updated homework 1
This is the real numberical solution to this problem. This is using the
radius as the dependent variable, which gives us the equation R(t+i) =
R(ti) -k (ti+1 - ti), instead of trying to use the original equation of
dv/dt = -kA. This gets rid of the unnessary cube root operation.